Adds cmake support for testing and benchmarking
Adds a dummy SYCL test and benchmark to ensure that compilation works.
Currently only tested on ComputeCpp, not TriSYCL.

Google benchmark does not seem to work with ComputeCpp when linked as a
library, but does when built as part of the project. The
SNN_DOWNLOAD_BENCHMARK is defaulted to ON to ensure that this happens
unless the user specifically wishes to prevent it.

To run the benchmarks use `ctest`.

To run the benchmarks use `ctest -C Benchmark -E test`.