Add window index helpers
Add helper functions to calculate the indices at which a pool or
convolution window starts for a given index. There are two
possibilities: computing an input window corresponding to an output
index, or computing an output window for an input index.

The functions return two indices. The window start index is the index in
the tensor which corresponds to the first value in the window, while the
filter start index is the index inside the window at which this value
occurs.

For a window of size 3, you might have the following:
```
in: 0 1 2
     \|/
out:  0
```
where the output index of 0 gives a window start of 0 and a filter start
of 0.

If the input index would be negative (because of some padding):
```
in: -1 0 1
      \|/
out:   0
```
then the function could return either `{window_start=-1,
filter_start=0}` or `{window_start=0, filter_start=1}`.

The filter start is particularly important for strides:
```
in: -1 0 1 2 3
      \|/ \|/
out:   0   1
```
The input index `1` is included in the window for both 0 and 1, so for
this the function would return `{window_start=0, filter_start=0}`.
However input `2` is only used by output 1 and only be the middle (index
1) value in the filter, so would return `{window_start=1,
filter_start=1}`.