Add ratio helper functions
Provide functions to round up integer division as well as a function to
round an integer up to the nearest multiple of another.

These functions only make sense for integer types, and have different
behaviours for positive and negative values, so the tests have to be
careful to explicitly test negative signed integers, as well as both
signed and unsigned integers for positive values.