Add fast integer division helpers
This uses the fast integer division technique outlined in "Division by
Invariant Integers using Multiplication" by Granlund and Montgomery
(http://dx.doi.org/10.1145/773473.178249), and the implementation is
based on that found in Chapter 10 (Figure 10-1) of "Hackers Delight" by
Warren.

The idea behind this fast division algorithm is to perform some
additional computations on the host to compute suitable magic numbers to
convert each division on the device into a multiply followed by a shift.