Create an snn_target function in SNNHelpers
Refactors the SNNHelpers cmake scripts to include a new snn_target
function which will set the common options for all targets and also
handle adding SYCL to any target which requires it. This will help with
adding library support later, as a library can be created with
`add_library(libname ...)` then configured and set up with
`snn_target(TARGET libname ...)`.