Add clang-formatting to Jenkins
This commit adds a clang-format file to the repository, and scripts
that will call the formatter on Jenkins to make sure that merge requests
are formatted correctly. There is a corresponding jenkins change to make
this happen automatically.

The clang-format version is unspecified by default, to match the unadorned
clang-format that comes with LLVM's binary archives. You can still specify
it when using it with the git hooks if you want to use it locally, say with
'clang-format-3.9', if that is what your package manager has called it.

One minor update has been made to the format file: instead of
inferring where * and & bind to, they are now specified to be on the
left.  Allowing it to be guessed by clang-format means that the style
was inconsistent across the repo's files. It is otherwise the same as
Google style from clang-format-3.9. The repository has been formatted
to match this style in this commit as well. From now on, there should
be no noisy format commits, ideally.