Include_after changes source not target properties
Each source file requires its own SYCL integration header, but does not
need every integration header included in the target. Instead of setting
the include flags on the target we instead should be setting the flags
on the source files.

Also changes how the compile order is switched when include after is
used, so that a target with multiple sources will not have them
overwritten by the SYCL file.