Fix finding different compute++ to libComputeCpp.so
If the user has some version of compute++ installed to a system path,
but passes in a different location for COMPUTECPP_PACKAGE_ROOT_DIR then
the runtime library found by cmake will be in the specified package
directory, whereas the device compiler will be in the system path.

Adding `NO_DEFAULT_PATH` to the find_program functions ensures that the
same paths are searched for both the runtime and the compiler.