Add tiled convolution implementation
Tiled convolutions implemented, along with tiled test cases.
Enabled the use of fast div. This might not actually lead to a
performance improvement as the register usage increases with it.
Fixed bugs when fast div is used when inappropriate.
Moved the 1D convolution functions into functor private methods.

This is a squashed and rebased commit of the original work done
by John, with small tweaks to make it work with the new CMake code.