Refactor how ComputeCpp is linked to SNN targets
Previously we had a number of options to work around the lack of support
for `target_link_libraries` with object libraries. A better workaround is
to avoid calling `target_link_libraries` and instead set the target
properties which would be set by it. This allows object library targets
to be treated the same as any other.