Adds object library helper to SNNHelpers
Adds `snn_object_library` helper function to the SNNHelpers cmake
module. This helps to avoid code repetition in the project's cmake
files and provides a single place to modify the configurations of all
object libraries.

This change does require a change in how the conv2d kernel sources are
added to the library target, as all the sources must be known before
calling snn_object_library to create the target rather than calling
`target_sources(...)`.