Add option to enable eigen exceptions
At the moment, if there's an async exception, it's lost, and the program
blows up. Eigen has a default handler that will catch them, but it
requires this macro. This new option is on by default. The death tests
now will fail if an exception is thrown when Eigen exceptions are
enabled.