Use generator expressions for compute++ flags
Rather than manually extracting the include directories and compile
definitions from a target, we can use generator expressions to allow
cmake to generate these for us. This is easier to maintain and has the
bonus of no longer requiring the include directories of a target are
completely set before calling `add_sycl_to_target`.