Add pooling tests
These tests are generated using the same framework as the convolution
tests. Only forward pooling is supported for now, as those are the only
operator types implemented.

The pooling tests for window11 and window 7, stride 4 fail when the
average operation is used, due to accuracy problems highlighted by the
size of the stride. For the moment, we just don't generate them.
The CMake outputs a warning about the fact it's skipping a test.