Increase precision in tests expected values
Instead of using the default conversion of a numpy array to a string, we
can use `numpy.array2string` and provide options to output a better
string representation. In particular we can specify that the values
should be output to a sufficient precision to ensure that the floating
point value is unique, and so we avoid any truncation or rounding of the
expected values.

By doing this we can re-enable the larger average pooling tests as they
no longer fail on lack of precision.