Add SNN backend with DeviceMemPointer
Add a backend that does not have any external dependencies. This helps
reduce the dependencies in our tests and benchmarks, as well as helping
to reduce compile times.

Changes the default backends used in tests and benchmarks to favour
SYCL-BLAS over Eigen, and to use our SNNBackend whenever there are no
matrix multiplies or performance is not a concern.