Fix basic benchmark loop behaviour
The basic benchmarks looped over a work submission, but did not wait.
That means the benchmark was benchmarking the length of time to submit
work to the SYCL queue, and nothing else. Now we wait on the event of
all the benchmarks so that we have an accurate idea of how long it
really takes.