Fix im2col pointer behaviour
The call to `matmul` in im2col relies on a `pointer<T>` being implicitly
convertible to `pointer<T const>`. This can cause problems for the
template type deduction if the conversion is not considered, as the
matmul assumes that the input pointers are all of `const` type. By
adding an explicit conversion when calling `matmul` we avoid any such
problems.

Also change the pointer set types to `internal_pointer_type` rather than
`pointer_type`, as these could potentially be different.