Fix exception handling
Since Eigen does not return events, a default event is constructed with
a null context and so an exception is thrown. A fix is achieved by first
checking that the SNNStatus is OK, and surrounding the wait_and_throw
with a try-catch block. Upon catching an exception, Google Benchmark is
notified that the benchmark failed and should skip to the next
benchmark.