Add tensor transpose operation
Add an operation to transpose multi-dimensional tensors which copies the
data in the input to the output permuting the dimensions in the tensors
according to a user provided permutation. The `i`-th output dimension will
be mapped to the `permutation[i]`-th input dimension.

Together with this general transpose operation, provide helpers to
convert between the well known tensor data formats NHWC and NCWH.