Specify maximum number of files to format at once
Even with `max-procs` specified, `xargs` may choose to only run one
`clang-format` command on all the project's files. By specifying the
`max-args` we limit the number of files that can be included in a single
`clang-format` call, and so force multiple instances to be called. This
ensures that the formatting is done in parallel.