Refactor selection of backend for benchmarking
Following from work done on specialising testing fixtures on backend
providers, this refactoring means the benchmarking fixtures make use of
the same code.

This means that benchmarks and tests can now use a single and consistent
interface for using a backend. Incidentally, this interface also
requires initialised device memory - meaning there is no longer the risk
of initial calculations on denorms or NaNs.