Replace placeholder accessors with MemObjects
Rather than using placeholder accessors on our interface, which are not
the best designed and certainly not designed for this purpose, we can
use a virtual class to abstract the buffer type. This MemObject class
wraps the buffer and provides methods to get accessors. In this way we
have much better control of our own API, and this is a first step
towards a better memory model in our public interface.