Provide vector support in workspace reduce
Our depthwise filter backprop kernel expects to be able to workspace
vectors. As each vector element corresponds to a single output element,
we cannot reduce the vector to a scalar, so the workspace reduction must
be able to reduce vectors. This requires fixing the indices used in the
workspace to ensure that no threads overwrite any others.