Provide benchmark params at runtime
Rather than generating many different types for the different benchmarks
with different parameters, generate the parameter sets at runtime using
the benchmark arguments.

The benchmark parameters are encoded using the
`benchmark_params::serialize()` function and can be decoded from the
`benchmark::State` using the corresponding
`benchmark_params::deserialize()` function. The serialized vector can be
passed to any Benchmark instance as an argument to run that benchmark
for the given parameters.