Add test for workgroup reduce function
The workgroup reduce function reduces a value across a workgroup,
returning a single value to the first work item in the workgroup.
Although used in the depthwise convolution kernels this did not
previously have any explicit tests. These tests should help isolate any
problems with the reduce function, as well as aiding any future
optimisations.

As part of this the flattened ID helpers are moved to their own helpers
file. These are used within the test and might be useful more generally.
Tests for these helpers are much harder to implement as the SYCL items
are not user constructible.