Generate MemObjects within Backends
Rather than extracting buffers and offsets from a backend and then
constructing a MemObject from these, it is easier for the backend to
control the whole process and provide the MemObject to the library. This
helps avoid any mistakes by a user where the wrong offset or extent is
used.

As a result of this change, the MemObjects need to provide access to
their internal data members. This was planned anyway, in order to allow
functions within the library to get access to the SYCL buffers, provided
that the BaseMemObject passed in can be dynamically cast to a known
MemObject.