Remove EIGEN_EXCEPTIONS option
This option causes complications in some of the tests which expect to be
able to use either Eigen or SYCL-BLAS, and with exceptions disabled
these libraries have different error reporting that make a generic test
much harder to write.

This option only affects the SYCL-DNN tests and benchmarks, with no
effect on the library itself. It makes little sense to disable
exceptions in our tests and benchmarks where we rely on them to handle
other errors.