Remove newline in exception error message
The error messages are used to set the benchmark skipped message, that
is encoded in the output benchmark json as is. This means that the
newline character is not escaped in the json and will cause parsing
errors. Changing to use a space may make the error a bit harder to read
for a human but will allow the json to be parsed.