Ensure Winograd minibatch is not larger than batch
The Winograd launcher splits up batches of images into smaller
minibatches if the workspace provided is not large enough. This scales
well for small workspace, but the current implementation will scale up
the number of minibatches used above the total number of input batches
if the workspace is larger than required. This can cause the transform
buffers to be much bigger than they need to be.