Check buffer size is valid in make_mem_object
The `make_mem_object` function is used throughout the library to create
MemObjects from SYCL buffers. This makes it a single point to check that
the size of the SYCL buffer is sufficient for whatever the library
needs. Adding an assert helps catch user errors where buffers that are
too small are passed into the library as GPUs can often not detect these
errors and will continue regardless, but overwrite data within the GPU
memory.