Add non deduced type parameters to helpers
Many functions in SYCL-DNN are templated, with most of the function
parameters being used as the same type. However there is rarely a requirement
that all the types are exactly the same, only that one type can be converted
to another, e.g. in the following it would be valid to divide a `long` by an
`int`, as the `int` can be easily be extended to `long`.

```
template <typename Index>
Index round_ratio_up(Index num, Index div);
```

Rather than having function templates that have a different function
parameter for each function parameter (and hence a new function declaration
for all pairs of types) we can ensure that the single template parameter is
used in a non-deducible context in the function parameters. This helps the
compile choose the right function instantiation and will automatically
convert the function parameters.

```
template <typename Index>
Index round_ratio_up(Index num, NonDeduced<Index>::type div);
```