Update matmul backend test for different backends
The matmul backend test makes a number of assumptions that are not valid
abut generic backends, such as the pointer type and that the internal
and external pointer types are the same. This is not always the case, so
the test needs to be generalised to ensure that it works correctly for
other backends.