Make SYCL-BLAS paths cache variables
If the SYCL-BLAS target is first created in one subdirectory, then
referenced again in a different subdirectory, the target will not be
available and so cmake will try to construct the target using
`find_package(SyclBLAS)`. Currently the path variables are not stored in
the cache, so are not set, but the SyclBLAS_Download target will exist.
This causes the second call to `find_package(SyclBLAS)` to fail as the
path variables are no longer available. To get around this, cache the
path variables so that they are available in all subdirectories.