Use SYCL-BLAS buffer iterators
Rather than using virtual pointers it is better to use the
`buffer_iterator` pointer like objects from SYCL-BLAS. These contain a
buffer and an offset in much the same way that the DeviceMemPtr objects
do.

This does require that the `buffer_iterator` behaves int he same way
as a pointer, so it must be default constructible and be implicitly
convertible to a `buffer_iterator<const T>`.