Set linker option /FORCE and compiler option /Zc:__cplusplus
The /FORCE option is required because SYCL-DNN relies on the compiler stripping unused code before linking. This is achieve on Linux with -Wl,--allow-shlib-undefined.

SYCL-DNN samples rely on Eigen which in turns rely on the macro __cplusplus to be the actual C++ version supported by the compiler. On MSVC this macro is not updated unless /Zc:__cplusplus is passed as a compiler option