Fix missing symbols from dead code
The functions that are called in the internal Tiled convolution launcher
depend on what type of convolution is used. As such many of the function
template instantiations in the original launch function could never be
called at runtime, however some compilers struggle to remove the calls
entirely.  This caused link errors when the linker tries to resolve
these functions that we never provide a definition for.

Previously we have worked around this with platform specific flags,
however it is better to change how the templates are instantiated to fix
the problem for all platforms.