Fix to build SYCL-DNN on Windows with MSVC (#5)
The object library targets need `sycl_dnn_EXPORTS` to be set to ensure that symbols are exported properly on windows. This is set automatically by cmake for normal library targets, but because SYCL-DNN uses object libraries this needs to be set in our build scripts.

SYCL-DNN samples rely on Eigen which in turns rely on the macro __cplusplus to be the actual C++ version supported by the compiler. On MSVC this macro is not updated unless /Zc:__cplusplus is passed as a compiler option, so we need to set this for all targets.