Add support for Ninja job pools
Some of the SYCL-DNN test targets contain a very large number of objects
that can use a large amount of memory when compiling. On systems with a
lot of CPU cores but limited memory this can cause the system to quickly
run out of memory. By providing an option to limit the number of
concurrent jobs to a user provided number this can be mitigated, or at
least worked around at the cost of longer compiler times.

Ninja supports job pools that can limit the number of concurrent jobs it
runs (see: https://ninja-build.org/manual.html#ref_pool), with some
support in CMake. We can add the test targets that use a lot of memory
to a special pool to restrict how many run at any one time.