Make backend name function static
Each backend is a different type without any virtual methods, so there
is no need for the `name()` function to be a member function rather than
a static function. This change allows better naming of test fixtures in
the future as an instance of a backend will not be required to get the
name.