Fix broken doxygen links
Doxygen expects to be run in the same directory that it can find the
files referred to in markdown files, and fails to find relative files
when it is run in a separate directory. We can change the cmake command
to use the main project directory as the working directory for the
Doxygen command, but still output to the build directory.

Some old unused options that are deprecated in newer version of
Doxygen are removed. Some redundant documentation is also removed.