Only benchmark Direct and Tiled with SNNBackend
The performance of these two algorithms shouldn't change with
the backend used. Particularly now that we offer a fourth backend,
we can reduce the amount of repeated benchmarks by only using the
SNNBackend to benchmark these implementations. A side effect is
that convolution benchmarks can now be built without any backends
enabled, meaning that only Direct and Tiled will run.