Add reductions as operations provided by backends
A reduction is a basic building block of many other operations which can
be optimised differently by different backends.
We introduce a generic 3D reduction that can be used both as an inner or
outer reduction. The middly dimension is always reduced.

This is still a naive implementation of a reduction but it is now
running on the device rather than the host. This changed the accuracy of
some tests (a 1 ULP difference in a mean reduction leads to a large
difference in Batchnorm).

This also fixes some undefined reference after testing with half enabled
and with other backends.