Use X Macros instead of manual instantiations
Previously, conv2d/CMakeLists.txt and launch_tiled.cc had to maintain
the same list of available tiled configurations. Instead, we now use
X Macros with a cmake generated 'tiled_variants.def'. This means that:

- CMakeLists.txt is the source of truth for available configs
- The list of available configs is automatically added to launch_tiled.cc

Downside is it makes launch_tiled_impl a *bit* illegible.

Another point worth bearing in mind: the order of the instantiations in
CMakeLists.txt is the priority list! In other words, the first
suitable configuration will be chosen!